user_microservice/pyproject.toml
[tool.poetry]
name = "discordapi"
version = "1.0"
description = "api похожая на дискорд"
authors = ["ledxDeliveryflopp <korstim18@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic-settings = "^2.6.1"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
passlib = "^1.7.4"
email-validator = "^2.2.0"
httpx = "^0.27.2"
uvicorn = "^0.32.0"
bcrypt = "^4.2.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pyjwt = "^2.10.0"
cryptography = "^43.0.3"
boto3 = "^1.35.65"
pyshorteners = "^1.0.1"
aiofiles = "^24.1.0"
python-multipart = "^0.0.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
markers = [
    "directly: Тестирование эндпоинта прямым запросом",
]